+++ R/type-rcrd.R
@@ -41,12 +41,6 @@ vec_restore.vctrs_rcrd <- function(x, to, ...) {
 }
 
 #' @export
-vec_proxy_equal.vctrs_rcrd <- function(x, ...) {
-  # Recursively proxy using a data frame
-  vec_proxy_equal(new_data_frame(x))
-}
-
-#' @export
 length.vctrs_rcrd <- function(x) {
   vec_size(x)
 }